Revert "format: Make pa_format_info_valid() stricter for PCM"
This reverts commit 12a202c510dcacbd2b85fcc1170484eb16fef491.

This is needed for now to avoid a clash in clients using json-glib. The
commit added a call to json_object_get_type() in client code that didn't
exist before, and this breaks some apps like Rhythmbox and Totem. This
will be fixed in the future by possibly dropping json-c as a dep.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=95135